// src/widgets/admin-machines/enums/category-id.ts
export enum ECategoryId {
    DIGITAL_PRINTING = '1',
    CATEGORY2 = '2',
    CATEGORY3 = '3',
    CATEGORY4 = '4',
    CATEGORY5 = '5',
    CATEGORY6 = '6',
    CATEGORY7 = '7',
    CATEGORY8 = '8',
    CATEGORY9 = '9',
    CATEGORY10 = '10',
    CATEGORY11 = '11',
    CATEGORY12 = '12',
    CATEGORY13 = '13',
    CATEGORY14 = '14',

}